Added width and height constants and refactored code
Now there are five constants which are used to specify most of the object's size parametres. I added these constants to easily change the size when needed.

I changed the small scores width to width / 25, so that the row layout can now have 8 small items with equal widths if it is not limited by height of the items.