FIxed android compatability
getting contents of support.txt was done through std::filesystem, however I added that file as a qt
resource and thus it was not possible to read the contents after installing the app as an apk(it was part of an executable). The temporary fix I came up with was just integrating the text inside the qml file as it is small and it is the only text file we need anyways.